More useful <meta> description
When people search QRL they currently see a list of the conferences the team has attended rather than a description of the network. Feel free to ignore the keywords if you have that managed in another section. I don't want to mess with SEO because it looks like you guys did a good job on that.