Update qip script with sanity checks
In order to keep the site up to date, sometimes it's required to create QIPS outside of using the script which may result in future incompatible URI's and/or the loss of QIPS in their entirety. It is strongly recommended to ensure there's no handmade qips that have not been merged, and, when they're merged, to ensure the URI is the same and wont be broken as the landing pages would have already been distributed through social media.